Misc cleanup and verify tests work.

+++ go/pkg/sieve/sieve_test.go
@@ -1,30 +1,29 @@
 package sieve
 
 import (
-	"math/big"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
+    "math/big"
+    "testing"
+    "github.com/stretchr/testify/assert"
 )
 
 func TestNthPrime(t *testing.T) {
-	sieve := NewSieve()
+    sieve := NewSieve()
 
-	assert.Equal(t, 2, sieve.NthPrime(0))
-	assert.Equal(t, 71, sieve.NthPrime(19))
-	assert.Equal(t, 541, sieve.NthPrime(99))
-	assert.Equal(t, 3581, sieve.NthPrime(500))
-	assert.Equal(t, 7793, sieve.NthPrime(986))
-	assert.Equal(t, 17393, sieve.NthPrime(2000))
-	assert.Equal(t, 15485867, sieve.NthPrime(1000000))
+    assert.Equal(t, int64(2), sieve.NthPrime(0))
+    assert.Equal(t, int64(71), sieve.NthPrime(19))
+    assert.Equal(t, int64(541), sieve.NthPrime(99))
+    assert.Equal(t, int64(3581), sieve.NthPrime(500))
+    assert.Equal(t, int64(7793), sieve.NthPrime(986))
+    assert.Equal(t, int64(17393), sieve.NthPrime(2000))
+    assert.Equal(t, int64(15485867), sieve.NthPrime(1000000))
 }
 
 func FuzzNthPrime(f *testing.F) {
-	sieve := NewSieve()
+    sieve := NewSieve()
 
-	f.Fuzz(func(t *testing.T, n int64) {
-		if !big.NewInt(sieve.NthPrime(n)).ProbablyPrime(0) {
-			t.Errorf("the sieve produced a non-prime number at index %d", n)
-		}
-	})
+    f.Fuzz(func(t *testing.T, n int64) {
+        if !big.NewInt(sieve.NthPrime(n)).ProbablyPrime(0) {
+            t.Errorf("the sieve produced a non-prime number at index %d", n)
+        }
+    })
 }
